+++ kit/config.js
@@ -86,11 +86,18 @@ if (SERVER) {
       // Create a set for routes -- to retrieve based on insertion order
       this.routes = new Set();
 
+      // Custom middleware
+      this.beforeMiddleware = new Set();
+      this.middleware = new Set();
+
       // Koa application function. But default, this is null
       this.koaAppFunc = null;
 
       // Flag for setting whether plain HTTP should be disabled
       this.enableHTTP = true;
+
+      // Flag for enabling the `Response-Time` Koa middleware timer
+      this.enableTiming = true;
 
       // Flag for enabling Koa Helmet HTTP hardening. True, by default.
       this.enableKoaHelmet = true;
@@ -102,9 +109,6 @@ if (SERVER) {
       // Options for enabling SSL. By default, this is null. If SSL is enabled
       // in userland, this would instead hold an object of options
       this.sslOptions = null;
-
-      // Custom middleware -- again, based on insertion order
-      this.middleware = new Set();
 
       // GraphQL schema (if we're using an internal server)
       this.graphQLSchema = null;
@@ -153,6 +157,11 @@ if (SERVER) {
       this.enableHTTP = false;
     }
 
+    // Disable timing the request. This will remove the `Response-Time` header
+    disableTiming() {
+      this.enableTiming = false;
+    }
+
     // Disable the optional `koa-bodyparser`, to prevent POST data being sent to
     // each request.  By default, body parsing is enabled.
     disableBodyParser() {
@@ -192,7 +201,12 @@ if (SERVER) {
       this.errorHandler = func;
     }
 
-    // Add custom middleware.  This should be an async func, for use with Koa
+    // Add custom middleware.  This should be an async func, for use with Koa.
+    // There are two entry points - 'before' and 'after'
+    addBeforeMiddleware(middlewareFunc) {
+      this.beforeMiddleware.add(middlewareFunc);
+    }
+
     addMiddleware(middlewareFunc) {
       this.middleware.add(middlewareFunc);
     }