+++ src/components/main/index.js
@@ -48,11 +48,11 @@ import { Redirect } from 'kit/lib/routing';
 // Child React components. Note:  We can either export one main React component
 // per file, or in the case of <Home>, <Page> and <WhenFound>, we can group
 // multiple components per file where it makes sense to do so
-import GraphQLMessage from 'src/components/graphql';
-import { Home, Page, WhenNotFound } from 'src/components/routes';
-import ReduxCounter from 'src/components/redux';
-import Stats from 'src/components/stats';
-import Styles from 'src/components/styles';
+import GraphQLMessage from 'components/graphql';
+import { Home, Page, WhenNotFound } from 'components/routes';
+import ReduxCounter from 'components/redux';
+import Stats from 'components/stats';
+import Styles from 'components/styles';
 
 // Styles
 import css from './main.scss';